st1 [1332][IMP] Adjust lot_id logic (#603)
