+++ account_invoice_line_quant/models/account_invoice_line.py
@@ -8,4 +8,4 @@ class AccountInvoiceLine(models.Model):
     _inherit = "account.invoice.line"
 
     quant_id = fields.Many2one("stock.quant", string="Stock Quant", copy=False)
-    lot_id = fields.Many2one(related="quant_id.lot_id", string="Case No.", store=True)
+    lot_id = fields.Many2one("stock.production.lot", string="Case No.", copy=False)
