--- account_invoice_line_quant/models/sale_order_line.py
@@ -1,16 +1,0 @@
-# Copyright 2019 Quartile Limited
-# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
-
-from odoo import api, models
-
-
-class SaleOrderLine(models.Model):
-    _inherit = "sale.order.line"
-
-    @api.multi
-    def _prepare_invoice_line(self, qty):
-        res = super(SaleOrderLine, self)._prepare_invoice_line(qty)
-        self.ensure_one()
-        if self.quant_id:
-            res.update({"quant_id": self.quant_id.id})
-        return res
