+++ account_invoice_line_quant/views/account_invoice_views.xml
@@ -1,18 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <record model="ir.ui.view" id="invoice_supplier_form">
-        <field name="name">account.invoice.supplier</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_supplier_form" />
-        <field name="arch" type="xml">
-            <xpath
-                expr="//field[@name='invoice_line_ids']/tree//field[@name='product_id']"
-                position="after"
-            >
-                <field name="lot_id" />
-            </xpath>
-        </field>
-    </record>
     <record model="ir.ui.view" id="invoice_form">
         <field name="name">account.invoice.form</field>
         <field name="model">account.invoice</field>
