+++ purchase_order_line_quant/models/__init__.py
@@ -1,4 +1,3 @@
-from . import account_invoice_line
 from . import account_invoice
 from . import purchase_order_line
 from . import purchase_order
