+++ purchase_order_line_quant/models/purchase_order_line.py
@@ -1,11 +1,11 @@
 # Copyright 2019 Quartile Limited
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 
-from odoo import fields, models
+from odoo import api, fields, models
 
 
 class PurchaseOrderLine(models.Model):
     _inherit = "purchase.order.line"
 
     quant_id = fields.Many2one("stock.quant", string="Stock Quant", copy=False)
-    lot_id = fields.Many2one(related="quant_id.lot_id", string="Case No.", store=True)
+    lot_id = fields.Many2one("stock.production.lot", string="Case No.", copy=False)
