+++ sale_order_line_quant/__manifest__.py
@@ -8,6 +8,7 @@
     "author": "Quartile Limited",
     "website": "https://www.quartile.co",
     "depends": [
+        "account_invoice_line_quant",
         "stock",
         "sale_margin",
         "stock_move_line_quant",
