+++ sale_order_line_quant/views/sale_order_views.xml
@@ -27,11 +27,7 @@
                                 ('usage', '=', 'internal')]"
                     attrs="{'readonly': [('is_mto','=',True)], 'required': [('is_mto','!=',True), ('tracking', '!=', 'none')]}"
                 />
-                <field
-                    name="lot_id"
-                    invisible="1"
-                    attrs="{'readonly': [('is_mto','=',True)]}"
-                />
+                <field name="lot_id" readonly="1" force_save="1" />
                 <field
                     name="route_id"
                     groups="sale_stock.group_route_so_lines"
@@ -61,11 +57,7 @@
                                 ('usage', '=', 'internal')]"
                     attrs="{'readonly': [('is_mto','=',True)], 'required': [('is_mto','!=',True), ('tracking', '!=', 'none')]}"
                 />
-                <field
-                    name="lot_id"
-                    invisible="1"
-                    attrs="{'readonly': [('is_mto','=',True)]}"
-                />
+                <field name="lot_id" readonly="1" force_save="1" />
                 <field
                     name="route_id"
                     groups="sale_stock.group_route_so_lines"
